+++ src/features/dashboard/timezone/index.ts
@@ -10,7 +10,6 @@ export type {
   ZonedDateTimeParts,
 } from './date-time'
 export {
-  createZonedTimeAxisLabelFormatter,
   formatDate,
   formatDateParts,
   formatDateRange,
